+++ src/main/java/com/aeropelican/productservice/entity/Product.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Data;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Table(name = "products")
 @Entity
@@ -12,7 +14,7 @@ public class Product {
     @Id
     @Column(name = "product_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer productId;
+    private Long productId;
     @Column(name = "product_name")
     private String productName;
     @Column(name = "description")
