+++ src/main/java/com/aeropelican/productservice/repository/ProductRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     boolean existsByProductNameIgnoreCase(String productName);
-    boolean existsByProductNameIgnoreCaseAndProductIdNot(String productName, Integer productId);
+    boolean existsByProductNameIgnoreCaseAndProductIdNot(String productName, Long productId);
 
 }
